requirements.txt
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
Authlib==1.6.8
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
cryptography==46.0.5
docstring_parser==0.17.0
exceptiongroup==1.3.1
fastapi==0.128.8
google-adk==1.18.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.139.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.40.1
google-cloud-bigtable==2.35.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.21.0
google-cloud-logging==3.13.0
google-cloud-monitoring==2.29.1
google-cloud-resource-manager==1.16.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.63.0
google-cloud-speech==2.37.0
google-cloud-storage==3.9.0
google-cloud-trace==1.18.0
google-crc32c==1.8.0
google-genai==1.47.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.2.5
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
kagglehub==0.3.13
Mako==1.3.10
MarkupSafe==3.0.3
mmh3==5.2.0
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==26.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.47
sqlalchemy-spanner==1.17.2
sqlparse==0.5.5
starlette==0.49.3
tenacity==9.1.2
tomli==2.4.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.39.0
watchdog==6.0.0
websockets==15.0.1
zipp==3.23.0
qdrant-client>=1.7.0
goodmem-adk>=0.1.0
